requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
typing_extensions>=4.0.0
lightgbm>=3.3.0
catboost>=1.1.0
xgboost>=1.7.0
autogluon.tabular>=1.0.0